backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas>=2.2.0
reportlab==4.0.7
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
supabase==2.0.2
pydantic==2.5.0
aiofiles==23.2.1 